pyproject.toml
[project]
name = "weather-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Aisaev"}
]
readme = "README.md"
requires-python = "3.12.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "aiohttp (>=3.11.18,<4.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
